pyproject.toml
[project]
name = "telegram-claude-bridge"
version = "0.1.0"
description = "Telegram ↔ Claude Code Permission Bridge"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "requests>=2.32.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.scripts]
telegram-claude-bridge = "bridge.daemon:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["bridge"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]
